requirements.txt
blinker==1.9.0
click==8.1.8
dot2tex==2.11.3
Flask==3.1.0
graphviz==0.20.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.2.1
pyparsing==3.2.3
Werkzeug==3.1.3
